Finally have generation and movement fixed. Still rooms overlapping though
