+++ src/main/kotlin/com/malcolmcrum/roguelike/entity/Entity.kt
@@ -1,5 +1,7 @@
 package com.malcolmcrum.roguelike.entity
 
-class Entity(var x: Int, var y: Int, val char: Char) {
+import com.malcolmcrum.roguelike.Point
 
+class Entity(var x: Int, var y: Int, val char: Char) {
+    constructor(point: Point, char: Char) : this(point.x, point.y, char)
 }
